+++ R/shinyapp.R
@@ -62,6 +62,7 @@ ui <- function(data, epsg) {
                                      br(),
                                      p("Application author: ",
                                        a("Josh Cullen", href = "https://joshcullen.github.io/"),
+                                       tags$br(),
                                        a("University of Florida", href = "http://www.ufl.edu"),
                                        style = "font-family: 'times'; font-si16pt")
                         ),
@@ -88,6 +89,7 @@ ui <- function(data, epsg) {
                           br(),
                           p("Application author: ",
                             a("Josh Cullen", href = "https://joshcullen.github.io/"),
+                            tags$br(),
                             a("University of Florida", href = "http://www.ufl.edu"),
                             style = "font-family: 'times'; font-si16pt")),