+++ tests/testthat/test-df_to_list.R
@@ -10,7 +10,8 @@ test_that("list is created properly from data frame", {
 
   #create data frame
   dat<- data.frame(id, date, dt, step, angle)
-  dat<- round_track_time(dat = dat, id = "id", int = 3600, tol = 15, time.zone = "UTC")
+  dat<- round_track_time(dat = dat, id = "id", int = 3600, tol = 15, time.zone = "UTC",
+                         units = "secs")
 
   #run function
   dat.list<- df_to_list(dat = dat, ind = "id")