+++ tests/testthat/test-round_track_time.R
@@ -13,7 +13,8 @@ test_that("in POSIXct format and only changed date and dt", {
   dat<- data.frame(id, date, dt, step, angle)
 
   #run function
-  dat1<- round_track_time(dat = dat, id = "id", int = 3600, tol = 20, time.zone = "UTC")
+  dat1<- round_track_time(dat = dat, id = "id", int = 3600, tol = 20, time.zone = "UTC",
+                          units = "secs")
 
   expect_is(dat1, "data.frame")
   expect_is(dat1$date, "POSIXct")