+++ tests/testthat/test-segment_behavior.R
@@ -16,7 +16,8 @@ test_that("segmentation model can effectively segment a dataset and return resul
 
   #create data frame and round time
   dat<- data.frame(id, date, dt, step, angle)
-  dat<- round_track_time(dat = dat, id = "id", int = 3600, tol = 15, time.zone = "UTC")
+  dat<- round_track_time(dat = dat, id = "id", int = 3600, tol = 15, time.zone = "UTC",
+                         units = "secs")
 
 
   #define limits for each bin