+++ src/RcppExports.cpp
@@ -77,6 +77,19 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// StoreZ
+IntegerMatrix StoreZ(IntegerVector z, IntegerMatrix store_z, int nobs);
+RcppExport SEXP _bayesmove_StoreZ(SEXP zSEXP, SEXP store_zSEXP, SEXP nobsSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< IntegerVector >::type z(zSEXP);
+    Rcpp::traits::input_parameter< IntegerMatrix >::type store_z(store_zSEXP);
+    Rcpp::traits::input_parameter< int >::type nobs(nobsSEXP);
+    rcpp_result_gen = Rcpp::wrap(StoreZ(z, store_z, nobs));
+    return rcpp_result_gen;
+END_RCPP
+}
 
 static const R_CallMethodDef CallEntries[] = {
     {"_bayesmove_rmultinom1", (DL_FUNC) &_bayesmove_rmultinom1, 2},
@@ -84,6 +97,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_bayesmove_SampleZAgg", (DL_FUNC) &_bayesmove_SampleZAgg, 7},
     {"_bayesmove_CumSumInv", (DL_FUNC) &_bayesmove_CumSumInv, 3},
     {"_bayesmove_SummarizeDat", (DL_FUNC) &_bayesmove_SummarizeDat, 5},
+    {"_bayesmove_StoreZ", (DL_FUNC) &_bayesmove_StoreZ, 3},
     {NULL, NULL, 0}
 };