+++ tests/testthat/test-cluster_obs.R
@@ -8,7 +8,7 @@ test_that("mixture model works", {
   tracks.list<- dplyr::bind_rows(tracks.list)
 
   #only retain id and discretized step length (SL) and turning angle (TA) columns
-  tracks<- subset(tracks.list, select = c(id, SL, TA))
+  tracks<- subset(tracks.list, select = c(SL, TA))
 
 
   set.seed(1)
@@ -23,10 +23,11 @@ test_that("mixture model works", {
                              nmaxclust = nmaxclust, nburn = nburn)
 
 
-  expect_length(dat.res, 5)
+  expect_length(dat.res, 6)
   expect_type(dat.res$loglikel, "double")
   expect_is(dat.res$theta, "matrix")
   expect_length(dat.res$phi, 2)
   expect_type(dat.res$phi, "list")
-  expect_length(dat.res$z, ngibbs)
+  expect_length(dat.res$z.MAP, nrow(tracks))
+  expect_equal(rowSums(dat.res$z.posterior), rep(nburn, nrow(tracks)))
 })