+++ NEWS.md
@@ -7,7 +7,8 @@
 * Export get_MAP_internal() function for use within workflow for observation-level behavior estimation using cluster_obs().
 * Update assign_tseg_internal() to account for non-consecutive `time1` variable.
 * Update plot_breakpoints_behav() to account for non-consecutive `time` variable.
-
+* Update some function arguments to be compatible with {furrr} package (>= 0.2.0).
+* Update prep_data() to also calculate net-squared displacement
 
 
 # bayesmove 0.1.0 (2020-09-21)